Добавлена обработка HTTP Status Code

+++ src/Board/Host/Board.Host.Api/Controllers/AnswerController.cs
@@ -2,6 +2,7 @@
 using Board.Application.AppData.Contexts.Comments.Services;
 using Board.Contracts.Answers;
 using Board.Contracts.Comments;
+using Board.Contracts.Posts;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Net;
@@ -15,7 +16,7 @@ namespace Board.Host.Api.Controllers;
 
 [ApiController]
 [Route(template: "answer")]
-[AllowAnonymous]
+//[AllowAnonymous]
 public class AnswerController : ControllerBase
 {
     private readonly IAnswerService _answerService;
@@ -34,15 +35,18 @@ public class AnswerController : ControllerBase
     /// <param name="dto"> Модель создания ответа. </param>
     /// <param name="cancellationToken"> Токен отмены операции. </param>
     /// <returns> Идентификаор созданного ответа. </returns>
+    ///<response code="401"> Пользователь не авторизован. </response>
+    ///<response code="201"> Ответ сохранён. </response>
     [Authorize]
     [HttpPost("CreateAnswer")]
+    [ProducesResponseType(StatusCodes.Status201Created)]
+    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
     public async Task<IActionResult> CreateByCommentIdAsync([FromBody] CreateAnswerDto dto, CancellationToken cancellationToken)
     {
         _logger.LogInformation("Добавление .");
 
         var result = await _answerService.CreateByCommentIdAsync(dto, cancellationToken);
         return StatusCode((int)HttpStatusCode.Created, result);
-        //return Ok();
     }
 
 
@@ -52,13 +56,28 @@ public class AnswerController : ControllerBase
     /// <param name="id"> Идентификатор ответа. </param>
     /// <param name="cancellationToken"> Токен отмены операции. </param>
     /// <returns> - </returns>
+    /// <response code="204"> Ответ успешно удалено. </response>
+    /// <response code="404"> Нет ответа с введенным идентификатором. </response>
+    /// <response code="403"> Нельзя удалить ответ другого пользователя. </response>
+    /// <response code="401"> Пользователь не авторизован. </response>
     [Authorize]
     [HttpDelete]
+    [ProducesResponseType(StatusCodes.Status204NoContent)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status403Forbidden)]
+    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
     public async Task<IActionResult> DeleteById(Guid id, CancellationToken cancellationToken)
     {
-        _logger.LogInformation("Добавление .");
+        _logger.LogInformation("Delete .");
+
+        var answerdto = await _answerService.GetByIdAsync(id, cancellationToken);
+        if (answerdto == null)
+            return StatusCode((int)HttpStatusCode.NotFound);
+        if (answerdto.AccId != _answerService.GetCurrentUserId() || _answerService.GetCurrentUserName() != "Admin")
+            return StatusCode((int)HttpStatusCode.Forbidden);
 
         await _answerService.DeleteById(id, cancellationToken);
+
         return StatusCode((int)HttpStatusCode.NoContent, null);
     }
 
@@ -69,12 +88,19 @@ public class AnswerController : ControllerBase
     /// <param name="commentId"> Идентификатор комментария. </param>
     /// <param name="cancellationToken"> Токен отмены операции. </param>
     /// <returns> Список ответов. </returns>
+    /// <response code="200"> Список ответов </response>
+    /// <response code="204"> Нет ответов или неверный идентификатор комментария. </response>
+    /// <response code="401"> Пользователь не авторизован. </response>
     [Authorize]
-    [HttpGet("GetAnswers/{CommentId}")]
+    [HttpGet("GetAnswers/{commentId}")]
+    [ProducesResponseType(StatusCodes.Status204NoContent)]
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
     public async Task<IActionResult> GetAnswersOnCommentsById(Guid commentId, CancellationToken cancellationToken)
     {
         //_logger.LogInformation("Получение списка комментариев к посту по идентификатору поста.");
         var result = await _answerService.GetAnswersOnCommentsById(commentId, cancellationToken);
+        if(result == null) return StatusCode((int)HttpStatusCode.NoContent);
         return await Task.Run(() => Ok(result));
     }
